requirements.txt
#requirements for the app to run these are all the libraries and imports needed to run the secure flask app
#it imports WTF forms which is secure form creation and input validation built in, SQLAlchemy and sqlite3 are used here
#as they are very secure. bcrypt used to hash the passwords for user accounts for extra security, and bleach used to sanitise user inputs on forms
#to prevent xss attacks
flask-sqlalchemy
requests
wtforms
limiter
Flask
Flask-WTF
Flask-Login
SQLAlchemy
Flask-Migrate
bcrypt
python-dotenv
cryptography
bleach
Flask-limiter
